Split long shortcuts transform test
